requirements.txt
git+https://github.com/EnokiUN/voltage
python-dotenv
Pillow
mariadb
